requirements.txt
dimod
dwave-system
fire
pyyaml
matplotlib